requirements.txt
# Core requirements
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
# psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-filter==23.5
psutil==5.9.8

# API Documentation
drf-spectacular==0.27.1

# Development tools
black==24.2.0
isort==5.12.0
flake8==7.0.0

# Testing
pytest==7.4.4
pytest-django==4.8.0

# Production
gunicorn==21.2.0 